% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{farmhash}
\alias{farmhash}
\title{Compute a 'FarmHash' digest for a vector of strings}
\usage{
farmhash(sv)
}
\arguments{
\item{sv}{Vector of strings for which a hash digest is requested}
}
\value{
A vector of \code{integer64} values (using the \pkg{bit64}
package to \sQuote{decode} the appropriate binary payload from a numeric
vector).
}
\description{
Compute a 'FarmHash' digest for a vector of strings
}
\examples{
if (requireNamespace("bit64", quietly=TRUE)) {
   suppressMessages(library(bit64))
}
farmhash(c("foo", "bar", "baz"))
}
