\name{RcppFarmHash-package}
\alias{RcppFarmHash-package}
\alias{RcppFarmHash}
\docType{package}
\title{\packageTitle{RcppFarmHash}}
\description{\packageDescription{RcppFarmHash}}
\section{Package Content}{\packageIndices{RcppFarmHash}}
\author{\packageAuthor{RcppFarmHash}}
\section{Maintainer}{\packageMaintainer{RcppFarmHash}}
\keyword{package}
